rename 'picker-view' to 'picklist'
