+++ src/components/ids-date-picker/demos/picker-only.html
@@ -21,7 +21,7 @@
         <ids-date-picker-popup target="#popup-trigger" trigger-type="click"></ids-date-picker-popup>
       </ids-layout-grid-cell>
       <ids-layout-grid-cell>
-        <ids-month-year-picker-view></ids-month-year-picker-view>
+        <ids-month-year-picklist></ids-month-year-picklist>
       </ids-layout-grid-cell>
     </ids-layout-grid>
   </ids-container>
